+++ src/mstring.h
@@ -29,8 +29,8 @@ private:
     friend class MStringArray;
 
     MStringData *fStr;
-    int fOffset;
-    int fCount;
+    size_t fOffset;
+    size_t fCount;
 
     void acquire() {
         ++fStr->fRefCount;
@@ -41,14 +41,14 @@ private:
         fStr = 0;
     }
     void destroy();
-    mstring(MStringData *fStr, int fOffset, int fCount);
-    void init(const char *str, int len);
+    mstring(MStringData *fStr, size_t fOffset, size_t fCount);
+    void init(const char *str, size_t len);
     const char *data() const { return fStr->fStr + fOffset; }
 public:
     mstring(const char *str);
     mstring(const char *str1, const char *str2);
     mstring(const char *str1, const char *str2, const char *str3);
-    mstring(const char *str, int len);
+    mstring(const char *str, size_t len);
 
     mstring(const class null_ref &):
         fStr(0),
@@ -70,7 +70,7 @@ public:
     }
     ~mstring();
 
-    int length() const { return fCount; }
+    size_t length() const { return fCount; }
     bool isEmpty() const { return 0 == fCount; }
     bool nonempty() const { return 0 < fCount; }
 
@@ -88,8 +88,8 @@ public:
 //    bool operator!=(const char *rv, size_t len) const { return !equals(rv, len); }
 
     mstring& operator=(const class null_ref &);
-    mstring substring(int pos) const;
-    mstring substring(int pos, int len) const;
+    mstring substring(size_t pos) const;
+    mstring substring(size_t pos, size_t len) const;
 
     int operator[](int pos) const { return charAt(pos); }
     int charAt(int pos) const;
@@ -98,7 +98,7 @@ public:
     int count(char ch) const;
 
     bool equals(const char *s) const;
-    bool equals(const char *s, int len) const;
+    bool equals(const char *s, unsigned len) const;
     bool equals(const mstring &s) const;
     int collate(const mstring &s, bool ignoreCase = false) const;
     int compareTo(const mstring &s) const;
